etl/requirements.txt
python-dotenv==1.0.0
psycopg2-binary~=2.9
pydantic==2.4.2
pydantic-settings==2.0.3
uvicorn>=0.15.0,<0.16.0
backoff==2.2.1
elastic-transport==8.4.0
elasticsearch==8.8.0